+++ changelog.md
@@ -5,6 +5,7 @@
 * Increased poison duration from Pufferfish Chowder
 * Item stacks are no longer limited to 99, thanks to InsaneLib, so now item stacks go up to 128
 * Naturally spawning guardians no longer spawn above a certain Y level and have less health and are smaller
+* Fixed unfair one-shot animation
 
 ## 5.5.1.2-alpha
 * Lowered iron armor regenerating absorption